Handlers are now represented in downstream objects by a ReedyHandler type defined in src/site-handlers/reedy-handler-type.ts
ReedyHandler is an interface/object with two methods;
1. getReedyElements which returns all the elements with text in them
2. getScrollableElement which returns an element for ReedyDirector to
   add a 'scroll' event listener to, which resets the current window
   when that element is scrolled
  ^^^^ This was done because an "interior scrollable" element is
  scrolled (a la the substack homepage)

Now every handler sub-script exports a ReedyHandler object implementing
the ReedyHandler interface;

If the site does not need (or does not have) a getScrollableElement
method written, a generic getScrollableElement method in the form of
() => undefined is appended to the object.

This method is used in the ReedyDirector constructor init process, in a ReedyDirector method called
setScrollableEventListener. If the getScrollableElement method returns
'undefined', ReedyDirector exits setScrollableEventListener before
setting any listener.